Setup Long mode paging. Triple faults :(

+++ Inc/typedefs.h
@@ -12,7 +12,7 @@ typedef BYTE BOOL;
 #define false 0
 // uint
 typedef unsigned int uint32_t;
-typedef unsigned long uint64_t;
+typedef unsigned long long uint64_t;
 typedef unsigned char uint8_t;
 typedef unsigned short int uint16_t;
 // pointers
